Add authentication and RBAC middleware
