+++ .github/workflows/ci.yml
@@ -14,6 +14,9 @@ jobs:
   lint-and-typecheck:
     name: Lint & Type Check
     runs-on: ubuntu-latest
+    env:
+      AUTH_SECRET: ${{ secrets.AUTH_SECRET }}
+      AUTH_USERS: ${{ secrets.AUTH_USERS }}
     steps:
       - name: Checkout
         uses: actions/checkout@v4
@@ -52,6 +55,9 @@ jobs:
     name: Build
     runs-on: ubuntu-latest
     needs: lint-and-typecheck
+    env:
+      AUTH_SECRET: ${{ secrets.AUTH_SECRET }}
+      AUTH_USERS: ${{ secrets.AUTH_USERS }}
     steps:
       - name: Checkout
         uses: actions/checkout@v4
@@ -80,6 +86,9 @@ jobs:
   security-audit:
     name: Security Audit
     runs-on: ubuntu-latest
+    env:
+      AUTH_SECRET: ${{ secrets.AUTH_SECRET }}
+      AUTH_USERS: ${{ secrets.AUTH_USERS }}
     steps:
       - name: Checkout
         uses: actions/checkout@v4
