+++ app/app/page.tsx
@@ -1,4 +1,5 @@
 import { redirect } from 'next/navigation';
+import { getServerAuthSession } from '@/app/lib/auth';
 
 // Metadata for the route
 export const metadata = {
@@ -10,6 +11,12 @@ export const metadata = {
  * App route - redirects to main dashboard.
  * This runs on the server so the redirect happens early.
  */
-export default function AppPage() {
+export default async function AppPage() {
+  const session = await getServerAuthSession();
+
+  if (!session) {
+    redirect('/auth/signin');
+  }
+
   redirect('/');
 }
