+++ package.json
@@ -46,6 +46,7 @@
     "@vanilla-extract/next-plugin": "^2.4.14",
     "@vanilla-extract/recipes": "^0.5.7",
     "@vanilla-extract/sprinkles": "^1.6.5",
+    "bcryptjs": "^3.0.2",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
@@ -59,6 +60,7 @@
     "lottie-web": "^5.13.0",
     "lucide-react": "^0.541.0",
     "next": "15.5.0",
+    "next-auth": "^4.24.11",
     "next-themes": "^0.4.6",
     "papaparse": "^5.5.3",
     "react": "19.1.0",
