+++ render.with-db.yaml
@@ -21,3 +21,7 @@ services:
           property: connectionString
       - key: NEXT_PUBLIC_API_URL
         sync: false
+      - key: AUTH_SECRET
+        sync: false
+      - key: AUTH_USERS
+        sync: false
