+++ render.yaml
@@ -13,6 +13,10 @@ services:
       # Optional; set via Render Dashboard if used by your deployment
       - key: NEXT_PUBLIC_API_URL
         sync: false
+      - key: AUTH_SECRET
+        sync: false
+      - key: AUTH_USERS
+        sync: false
       # Skip git hooks in CI installs
       - key: HUSKY
         value: "0"
